Separate origin from permissionRequest, move activeChainID querying to internal-ethereum-provider
