+++ background/services/internal-ethereum-provider/db.ts
@@ -1,10 +1,11 @@
 import Dexie from "dexie"
 import { TALLY_INTERNAL_ORIGIN } from "./constants"
 
-type ActiveChainId = {
+export type ActiveChainId = {
   chainId: string
   origin: string
 }
+
 export class InternalEthereumProviderDatabase extends Dexie {
   private activeChainId!: Dexie.Table<ActiveChainId, string>
 
@@ -22,12 +23,6 @@ export class InternalEthereumProviderDatabase extends Dexie {
     })
   }
 
-  private async getInternalActiveChain(): Promise<ActiveChainId> {
-    return this.activeChainId.get({
-      origin: TALLY_INTERNAL_ORIGIN,
-    }) as Promise<ActiveChainId>
-  }
-
   async setActiveChainIdForOrigin(
     chainId: string,
     origin: string
@@ -35,15 +30,10 @@ export class InternalEthereumProviderDatabase extends Dexie {
     return this.activeChainId.put({ origin, chainId })
   }
 
-  async getActiveChainIdForOrigin(origin: string): Promise<string> {
-    const activeChainId = await this.activeChainId.get({ origin })
-    if (!activeChainId) {
-      // If this is a new dapp or the dapp has not implemented wallet_switchEthereumChain
-      // use the default network.
-      const defaultChainId = (await this.getInternalActiveChain()).chainId
-      return defaultChainId
-    }
-    return activeChainId?.chainId
+  async getActiveChainIdForOrigin(
+    origin: string
+  ): Promise<ActiveChainId | undefined> {
+    return this.activeChainId.get({ origin })
   }
 }
